backend/requirements.txt
fastapi
uvicorn
qdrant-client
python-dotenv
langchain-text-splitters
langchain-openai
openai